+++ src/Http/ClientInterface.php
@@ -76,4 +76,18 @@ interface ClientInterface
      * @param bool $silently To ask client to not throw the exception (execute throw $throwable) or not. Default false
      */
     public function errorInRequest(Throwable $throwable, bool $silently = false): ClientInterface;
+
+    /**
+     * To set the client to be not silent, by default, when it will send a missing response and throw an exception
+     *
+     * @return ClientInterface
+     */
+    public function mustSendAResponse(): ClientInterface;
+
+    /**
+     * To set the client to be silent, by default, when it will send a missing response and throw an exception
+     *
+     * @return ClientInterface
+     */
+    public function sendAResponseIsOptional(): ClientInterface;
 }